Add files via upload
a simple PHP application that will ask users to enter Name, E-Mail, and Phone Number values , and when the submit button is clicked, will then use the data from the text fields to add more lines at the end of the message. Make sure you add appropriate input validations for all the fields.